+++ mongo_libs.py
@@ -196,6 +196,55 @@ def create_instance(cfg_file, dir_path, class_name):
         auth=cfg.auth, conf_file=cfg.conf_file, **config)
 
 
+def create_security_config(**kwargs):
+
+    """Function:  create_security_config
+
+    Description:  Create security configuration object.  Can take a
+        configuration file and load it or a pre-loaded configuration module or
+        a mongo instance.
+
+    Arguments:
+        (input) **kwargs:
+            cfg_file -> Configuration file name
+            dir_path -> Directory path to configuration file
+            cfg -> Mongo instance or loaded configuration module
+        (output) config -> Dictionary of security config parameters
+
+    """
+
+    config = dict()
+    cfg_file = kwargs.get("cfg_file", None)
+    dir_path = kwargs.get("dir_path", None)
+    cfg = kwargs.get("cfg", None)
+
+    if  cfg_file and dir_path:
+        cfg = gen_libs.load_module(cfg_file, dir_path)
+
+    if cfg:
+        config["auth_db"] = cfg.auth_db if hasattr(cfg, "auth_db") else "admin"
+        config["auth_mech"] = cfg.auth_mech if hasattr(
+            cfg, "auth_mech") else "SCRAM-SHA-1"
+        config["ssl_client_ca"] = cfg.ssl_client_ca if hasattr(
+            cfg, "ssl_client_ca") else None
+        config["ssl_client_cert"] = cfg.ssl_client_cert if hasattr(
+            cfg, "ssl_client_cert") else None
+        config["ssl_client_key"] = cfg.ssl_client_key if hasattr(
+            cfg, "ssl_client_key") else None
+        config["ssl_client_phrase"] = cfg.ssl_client_phrase if hasattr(
+            cfg, "ssl_client_phrase") else None
+        config["auth_type"] = cfg.auth_type if hasattr(
+            cfg, "auth_type") else None
+        config["tls_ca_certs"] = cfg.tls_ca_certs if hasattr(
+            cfg, "tls_ca_certs") else None
+        config["tls_certkey"] = cfg.tls_certkey if hasattr(
+            cfg, "tls_certkey") else None
+        config["tls_certkey_phrase"] = cfg.tls_certkey_phrase if hasattr(
+            cfg, "tls_certkey_phrase") else None
+
+    return config
+
+
 def crt_base_cmd(mongo, prog_name, **kwargs):
 
     """Function:  crt_base_cmd