+++ test/unit/mongo_libs/unit_test_run.sh
@@ -9,6 +9,7 @@ echo "Unit test: mongo_class"
 /usr/bin/python ./test/unit/mongo_libs/add_tls_cmd.py
 /usr/bin/python ./test/unit/mongo_libs/create_cmd.py
 /usr/bin/python ./test/unit/mongo_libs/create_instance.py
+/usr/bin/python ./test/unit/mongo_libs/create_security_config.py
 /usr/bin/python ./test/unit/mongo_libs/crt_base_cmd.py
 /usr/bin/python ./test/unit/mongo_libs/crt_coll_inst.py
 /usr/bin/python ./test/unit/mongo_libs/data_out.py